Remove CollectionSpecificDao, Create a common Api, Add Device API Files
Removed CollectionSpecificDao.java because it has only a single
instance of a cache, which creates a(n obvious) problem when the
ObjectId is the same (in the case of all modules where the parentId
is same, i.e. Device, Event, User).
Defined individual static caches for all prior subclasses of
CollectionSpecificDao for now.

Add created & updated ZonedDateTime properties for all Models.
Serialization / Deserialization is still a WIP.